+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/handler/pusher/DrainPusher.java
@@ -1,0 +1,30 @@
+package com.dwarfeng.settingrepo.impl.handler.pusher;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * 简单的丢弃掉所有信息的推送器。
+ *
+ * @author DwArFeng
+ * @since 1.1.2
+ */
+@Component
+public class DrainPusher extends AbstractPusher {
+
+    public static final String SUPPORT_TYPE = "drain";
+
+    public DrainPusher() {
+        super(SUPPORT_TYPE);
+    }
+
+    @Override
+    public void formatReset() {
+    }
+
+    @Override
+    public String toString() {
+        return "DrainPusher{" +
+                "pusherType='" + pusherType + '\'' +
+                '}';
+    }
+}